+++ src/main/java/com/example/webshop/exception/apiException/badRequestException/CategoryNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.webshop.exception.apiException.badRequestException;
 
 public class CategoryNotFoundException extends BadRequestException{
-	public CategoryNotFoundException(Integer id){
+	public CategoryNotFoundException(Long id){
 		super("Category with id " + id + " not found");
 	}
