+++ src/test/java/com/example/webshop/controller/ProductControllerTest.java
@@ -24,7 +24,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
 @ExtendWith(MockitoExtension.class)
-class ProductControllerTest {
+class ProductControllerTest{
 
 	private MockMvc mockMvc;
 
@@ -35,107 +35,95 @@ class ProductControllerTest {
 	private ProductController productController;
 
 	@BeforeEach
-	void setUp() {
+	void setUp(){
 		mockMvc = MockMvcBuilders.standaloneSetup(productController).build();
 	}
 
 	@Test
-	void getAllCategoriesTest() throws Exception {
-		CategoryDto categoryDto = new CategoryDto(1, "Electronics", "Electronic Items");
+	void getAllCategoriesTest() throws Exception{
+		CategoryDto categoryDto = new CategoryDto(1L, "Electronics", "Electronic Items");
 		given(productService.getAllCategories()).willReturn(Collections.singletonList(categoryDto));
 
-		mockMvc
-			.perform(get("/products/category")
-			.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(status().isOk())
-			.andExpect(content().contentType(MediaType.APPLICATION_JSON));
+		mockMvc.perform(get("/products/category").contentType(MediaType.APPLICATION_JSON)).andExpect(status().isOk())
+		       .andExpect(content().contentType(MediaType.APPLICATION_JSON));
 	}
 
 	@Test
-	void createCategory_Success() throws Exception {
+	void createCategory_Success() throws Exception{
 		CategoryDto categoryDto = new CategoryDto(null, "Electronics", "Electronic Items");
-		CategoryDto savedCategoryDto = new CategoryDto(1, "Electronics", "Electronic Items");
+		CategoryDto savedCategoryDto = new CategoryDto(1L, "Electronics", "Electronic Items");
 
 		given(productService.createCategory(any(CategoryDto.class))).willReturn(savedCategoryDto);
 
-		mockMvc
-			.perform(post("/products/category")
-			.contentType(MediaType.APPLICATION_JSON)
-			.content(new ObjectMapper().writeValueAsString(categoryDto)))
-			.andExpect(status().isCreated())
-			.andExpect(content().json(new ObjectMapper().writeValueAsString(savedCategoryDto)));
+		mockMvc.perform(post("/products/category").contentType(MediaType.APPLICATION_JSON)
+		                                          .content(new ObjectMapper().writeValueAsString(categoryDto))).andExpect(
+			status().isCreated()).andExpect(content().json(new ObjectMapper().writeValueAsString(savedCategoryDto)));
 	}
 
 	@Test
-	void deleteCategory_Success() throws Exception {
-		doNothing().when(productService).deleteCategory(1);
+	void deleteCategory_Success() throws Exception{
+		doNothing().when(productService).deleteCategory(1L);
 
-		mockMvc
-			.perform(delete("/products/category/1")
-			.contentType(MediaType.APPLICATION_JSON))
-			.andExpect(status().isNoContent());
+		mockMvc.perform(delete("/products/category/1").contentType(MediaType.APPLICATION_JSON)).andExpect(
+			status().isNoContent());
 
-		verify(productService, times(1)).deleteCategory(1);
+		verify(productService, times(1)).deleteCategory(1L);
 	}
 
 	@Test
-	void getAllProductsTest() throws Exception {
-		ProductDTO productDto = new ProductDTO(); // Initialize your ProductDTO
+	void getAllProductsTest() throws Exception{
+		ProductDTO productDto = new ProductDTO();
 		// Mock service method
 		given(productService.getAllProducts()).willReturn(Collections.singletonList(productDto));
 
-		mockMvc.perform(get("/products")
-			                .contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isOk())
-		       .andExpect(jsonPath("$", hasSize(1)));
+		mockMvc.perform(get("/products").contentType(MediaType.APPLICATION_JSON)).andExpect(status().isOk()).andExpect(
+			jsonPath("$", hasSize(1)));
 	}
 
 	@Test
-	void createProductTest() throws Exception {
-		ProductDTO productDto = new ProductDTO(); // Initialize with test data
-		ProductDTO savedProductDto = new ProductDTO(); // Initialize with expected data
+	void createProductTest() throws Exception{
+		ProductDTO productDto = new ProductDTO();
+		ProductDTO savedProductDto = new ProductDTO();
 
 		given(productService.createProduct(any(ProductDTO.class))).willReturn(savedProductDto);
 
-		mockMvc.perform(post("/products")
-			                .contentType(MediaType.APPLICATION_JSON)
-			                .content(new ObjectMapper().writeValueAsString(productDto)))
-		       .andExpect(status().isCreated());
+		mockMvc.perform(post("/products").contentType(MediaType.APPLICATION_JSON)
+		                                 .content(new ObjectMapper().writeValueAsString(productDto))).andExpect(
+			status().isCreated());
 	}
 
 	@Test
-	void getProductByIdTest() throws Exception {
-		Integer productId = 1;
-		ProductDTO productDto = new ProductDTO(); // Initialize with expected data
+	void getProductByIdTest() throws Exception{
+		Long productId = 1L;
+		ProductDTO productDto = new ProductDTO();
 
 		given(productService.getProductById(productId)).willReturn(productDto);
 
-		mockMvc.perform(get("/products/{productId}", productId).contentType(MediaType.APPLICATION_JSON))
-	       .andExpect(status().isOk());
+		mockMvc.perform(get("/products/{productId}", productId).contentType(MediaType.APPLICATION_JSON)).andExpect(
+			status().isOk());
 	}
 
 	@Test
-	void updateProductTest() throws Exception {
-		ProductDTO productDto = new ProductDTO(); // Initialize with test data
-		ProductDTO updatedProductDto = new ProductDTO(); // Initialize with expected updated data
+	void updateProductTest() throws Exception{
+		ProductDTO productDto = new ProductDTO();
+		ProductDTO updatedProductDto = new ProductDTO();
 
 		given(productService.updateProduct(any(ProductDTO.class))).willReturn(updatedProductDto);
 
-		mockMvc.perform(put("/products")
-			                .contentType(MediaType.APPLICATION_JSON)
-			                .content(new ObjectMapper().writeValueAsString(productDto)))
-		       .andExpect(status().isOk());
+		mockMvc.perform(put("/products").contentType(MediaType.APPLICATION_JSON)
+		                                .content(new ObjectMapper().writeValueAsString(productDto))).andExpect(
+			status().isOk());
 	}
 
 	@Test
-	void deleteProductTest() throws Exception {
-		Integer productId = 1;
+	void deleteProductTest() throws Exception{
+		Long productId = 1L;
 		doNothing().when(productService).deleteProduct(productId);
 
-		mockMvc.perform(delete("/products/{productId}", productId)
-			                .contentType(MediaType.APPLICATION_JSON))
-		       .andExpect(status().isNoContent());
+		mockMvc.perform(delete("/products/{productId}", productId).contentType(MediaType.APPLICATION_JSON)).andExpect(
+			status().isNoContent());
 
 		verify(productService, times(1)).deleteProduct(productId);
 	}
+
 }
